requirements.txt
Flask
feedparser
requests
google-auth
beautifulsoup4
